Separated some code. Landing page layout fix.
